Move Profile out of morph and implement into UnpackConfigurations
